Setting self.context.done = True before calling the callback in an async handler.
This should fix the issue when there is an error in the callback,
which raises an aync _error (async.py ln 627), which in turn calls the callback
again since the handler context does not appear complete (even though
we alrady read the http_content). These appear to always trigger a
socket close with 'premature close' (async.py ln 664).